fix header h-full
