+++ frontend/src/components/shared/Header.tsx
@@ -42,8 +42,11 @@ export function Header({ username, avatarImage }: HeaderProps) {
 
   return (
     <header>
-      <Navbar fluid className="h-[3rem] md:h-[4rem] px-3 md:px-8 border-b border-gray-200 bg-white">
-        <div className="h-full flex items-center justify-between w-full max-w-6xl mx-auto">
+      <Navbar
+        fluid
+        className="h-[3rem] md:h-[4rem] px-3 md:px-8 border-b border-gray-200 bg-white [&>div]:h-full"
+      >
+        <div className="flex items-center justify-between w-full max-w-6xl mx-auto">
           {/* logo */}
           <div className="flex items-center gap-10">
             <NavbarBrand as={Link} href="/">
